Move SongSource to it's own file
