+++ lib/music_player/music_player.dart
@@ -15,6 +15,7 @@ import 'package:musbx/music_player/slowdowner/slowdowner.dart';
 import 'package:musbx/music_player/song.dart';
 import 'package:musbx/music_player/song_history.dart';
 import 'package:musbx/music_player/song_preferences.dart';
+import 'package:musbx/music_player/song_source.dart';
 import 'package:musbx/widgets.dart';
 
 /// The state of [MusicPlayer].
